# gen/event/event_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "event/event_api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "EventAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/events/types": {
      "get": {
        "summary": "ListAuditEventTypes returns the list of audit events using filters.",
        "operationId": "EventAPI_ListAuditEventTypes",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAuditEventTypesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "EventAPI"
        ]
      }
    },
    "/v1/events:search": {
      "post": {
        "summary": "ListAuditEvents returns the list of audit events using filters.",
        "operationId": "EventAPI_ListAuditEvents",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAuditEventsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "ListAuditEventsRequest to view events that were previously saved.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListAuditEventsRequest"
            }
          }
        ],
        "tags": [
          "EventAPI"
        ]
      }
    },
    "/v1/hooks/events": {
      "post": {
        "summary": "Creates an event. The method requires Hook authentication, it's optimized for hooks to respond quickly.",
        "operationId": "EventAPI_CreateHookEvent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateHookEventResponse"
            },
            "examples": {
              "application/json": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "CreateHookEventRequest to be used to store okta event in persistent storage.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateHookEventRequest"
            }
          }
        ],
        "tags": [
          "Events management"
        ],
        "produces": [
          "application/json"
        ]
      }
    }
  },
  "definitions": {
    "ActionStatusMessageActionStatus": {
      "type": "object",
      "properties": {
        "code": {
          "$ref": "#/definitions/ActionStatusMessageActionStatusCode",
          "description": "Code assigned to this status."
        },
        "message": {
          "type": "string",
          "description": "Message to elaborate the reason of action failure or details of successful completion."
        }
      },
      "description": "ActionStatus represents a result of an action."
    },
    "ActionStatusMessageActionStatusCode": {
      "type": "string",
      "enum": [
        "UNSPECIFIED",
        "SUCCESS",
        "FAILURE"
      ],
      "default": "UNSPECIFIED",
      "description": "ActionStatusCode codifies result of an action.\n\n - UNSPECIFIED: UNSPECIFIED for non-passed value or ambiguous result of the operation.\n - SUCCESS: SUCCESS represents SUCCESS action status.\n - FAILURE: FAILURE represents FAILURE action status."
    },
    "ActionTypeMessageActionType": {
      "type": "string",
      "enum": [
        "UNSPECIFIED",
        "CREATE",
        "UPDATE",
        "DELETE"
      ],
      "default": "UNSPECIFIED",
      "description": "ActionType represents audit events' action types.\n\n - UNSPECIFIED: UNSPECIFIED represents unknown action type.\n - CREATE: CREATE represents action type = CREATE.\n - UPDATE: UPDATE represents action type = UPDATE.\n - DELETE: DELETE represents action type = DELETE."
    },
    "ClientUserAgent": {
      "type": "object",
      "properties": {
        "os": {
          "type": "string"
        },
        "browser": {
          "type": "string"
        },
        "rawUserAgent": {
          "type": "string"
        }
      }
    },
    "EventActor": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "alternateId": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        }
      }
    },
    "EventAuthContext": {
      "type": "object",
      "properties": {
        "authenticationContext": {
          "type": "string"
        },
        "externalSessionId": {
          "type": "string"
        }
      }
    },
    "EventClient": {
      "type": "object",
      "properties": {
        "zone": {
          "type": "string"
        },
        "device": {
          "type": "string"
        },
        "ipAddress": {
          "type": "string"
        },
        "userAgent": {
          "$ref": "#/definitions/ClientUserAgent"
        }
      }
    },
    "EventOutcome": {
      "type": "object",
      "properties": {
        "result": {
          "type": "string"
        },
        "reason": {
          "type": "string"
        }
      }
    },
    "EventTransaction": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "IdPEventDataEvent": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        },
        "severity": {
          "type": "string"
        },
        "client": {
          "$ref": "#/definitions/EventClient"
        },
        "actor": {
          "$ref": "#/definitions/EventActor"
        },
        "outcome": {
          "$ref": "#/definitions/EventOutcome"
        },
        "uuid": {
          "type": "string"
        },
        "published": {
          "type": "string"
        },
        "eventType": {
          "type": "string"
        },
        "displayMessage": {
          "type": "string"
        },
        "transaction": {
          "$ref": "#/definitions/EventTransaction"
        },
        "authenticationContext": {
          "$ref": "#/definitions/EventAuthContext"
        }
      }
    },
    "IdpEventIdPEventData": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/IdPEventDataEvent"
          }
        }
      }
    },
    "fspField": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "Name of the field (name of the column when applicable to SQL).\nIf a filter exists then the name field must not be empty since that\ncan result in an invalid SQL query and lead to non-descriptive errors."
        },
        "value": {
          "description": "Value of the field."
        }
      },
      "description": "Field of an object."
    },
    "fspFilter": {
      "type": "object",
      "properties": {
        "filterType": {
          "$ref": "#/definitions/fspFilterType",
          "description": "filter_type defines logic of the filter."
        },
        "field": {
          "$ref": "#/definitions/fspField",
          "description": "field defines based on which data filtering must occur."
        }
      },
      "description": "Filter is used to filter results for endpoints that return list of values."
    },
    "fspFilterType": {
      "type": "string",
      "enum": [
        "EQUAL",
        "NOT_EQUAL",
        "GREATER",
        "GREATER_OR_EQUAL",
        "LESS",
        "LESS_OR_EQUAL"
      ],
      "default": "EQUAL",
      "description": "FilterType represents type of filter that will be applied to SQL query.\n\n - EQUAL: In case single value is provided, EQUAL will trigger \"=\" SQL behaviour.\nIf array of values provided, EQUAL will trigger \"IN\" SQL behaviour.\n - NOT_EQUAL: In case single value is provided, NOT EQUAL will trigger \"\u003c\u003e\" SQL behaviour.\nIf array of values provided, EQUAL will trigger \"NOT IN\" SQL behaviour.\n - GREATER: GREATER is mapped to SQL '\u003e'.\n - GREATER_OR_EQUAL: GREATER_OR_EQUAL is mapped to SQL '\u003e='.\n - LESS: LESS is mapped to SQL '\u003c'.\n - LESS_OR_EQUAL: LESS_OR_EQUAL is mapped to SQL '\u003c='."
    },
    "fspFilteringSortingPagination": {
      "type": "object",
      "properties": {
        "filters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/fspFilter"
          },
          "description": "filters are optional. No filters will be applied if absent. If multiple provided - will be joined using AND logic."
        },
        "sortingParams": {
          "$ref": "#/definitions/fspSortingParams",
          "description": "sorting_params are optional. No sorting will be applied if absent."
        },
        "pageParams": {
          "$ref": "#/definitions/fspPageParams",
          "description": "page_params are optional. No pagination will be applied if absent."
        }
      },
      "description": "FilteringSortingPagination aggregates params for filtering, sorting and pagination."
    },
    "fspPageParams": {
      "type": "object",
      "properties": {
        "pageSize": {
          "type": "integer",
          "format": "int64",
          "description": "Maximum number of results per page."
        },
        "index": {
          "type": "integer",
          "format": "int64",
          "description": "Index of the requested page, starts from 0."
        }
      },
      "description": "PageParams represents page request parameters for pagination."
    },
    "fspPageTotals": {
      "type": "object",
      "properties": {
        "totalItems": {
          "type": "integer",
          "format": "int64",
          "description": "Total number of results."
        },
        "totalPages": {
          "type": "integer",
          "format": "int64",
          "description": "Total number of pages."
        }
      },
      "description": "PageTotals represents total values for pagination and are included in paginated responses."
    },
    "fspSortingOrder": {
      "type": "string",
      "enum": [
        "ASC",
        "DESC"
      ],
      "default": "ASC",
      "description": " - ASC: ASC is for ascending sorting order.\n - DESC: DESC is for descending sorting order."
    },
    "fspSortingParams": {
      "type": "object",
      "properties": {
        "fieldName": {
          "type": "string"
        },
        "order": {
          "$ref": "#/definitions/fspSortingOrder"
        }
      },
      "description": "SortingParams are used to sort results for endpoints that return list of values."
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n\nExample 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\nExample 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\n\nJSON\n\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "protobufNullValue": {
      "type": "string",
      "enum": [
        "NULL_VALUE"
      ],
      "default": "NULL_VALUE",
      "description": "`NullValue` is a singleton enumeration to represent the null value for the\n`Value` type union.\n\n The JSON representation for `NullValue` is JSON `null`.\n\n - NULL_VALUE: Null value."
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AuditEvent": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Audit event ID."
        },
        "oktaUserId": {
          "type": "string",
          "description": "User's okta ID is an ID that user has in Okta."
        },
        "requestId": {
          "type": "string",
          "description": "Unique ID assigned to each request."
        },
        "userIpAddress": {
          "type": "string",
          "description": "User's remote IP address."
        },
        "userAgent": {
          "type": "string",
          "description": "User agent of user's browser."
        },
        "userRole": {
          "type": "string",
          "description": "User's role in Portal."
        },
        "organizationId": {
          "type": "string",
          "description": "Organization ID that affected object belongs to."
        },
        "actionType": {
          "$ref": "#/definitions/ActionTypeMessageActionType",
          "title": "Action type"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time",
          "description": "Timestamp when the event occurred."
        },
        "eventSource": {
          "type": "string",
          "description": "The source module/service that triggered audit event (might be equal to account_id for events not related to user action)."
        },
        "actionStatus": {
          "$ref": "#/definitions/ActionStatusMessageActionStatus",
          "description": "The status of action execution."
        },
        "objects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1AuditObject"
          },
          "description": "The information about affected objects in the system."
        }
      },
      "description": "Event represents full audit event."
    },
    "v1AuditObject": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Object ID."
        },
        "name": {
          "type": "string",
          "description": "Object name."
        },
        "type": {
          "type": "string",
          "description": "Object type."
        },
        "initialState": {
          "type": "object",
          "description": "Object's initial attribute states that have been affected by user."
        },
        "resultState": {
          "type": "object",
          "description": "Object's resulted attribute states that have been affected by user."
        }
      },
      "description": "AuditObject represents object with it's attributes that is affected by user action."
    },
    "v1CreateAuditEventResponse": {
      "type": "object"
    },
    "v1CreateHookEventRequest": {
      "type": "object",
      "properties": {
        "event": {
          "$ref": "#/definitions/v1IdpEvent"
        }
      },
      "description": "CreateHookEventRequest to be used to store okta event in persistent storage."
    },
    "v1CreateHookEventResponse": {
      "type": "object"
    },
    "v1IdpEvent": {
      "type": "object",
      "properties": {
        "eventType": {
          "type": "string"
        },
        "eventTypeVersion": {
          "type": "string"
        },
        "cloudEventsVersion": {
          "type": "string"
        },
        "contentType": {
          "type": "string"
        },
        "eventId": {
          "type": "string"
        },
        "eventTime": {
          "type": "string"
        },
        "source": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/IdpEventIdPEventData"
        }
      },
      "title": "Okta event example https://developer.okta.com/docs/concepts/event-hooks/#sample-event-delivery-payload"
    },
    "v1ListAuditEventTypesResponse": {
      "type": "object",
      "properties": {
        "actionTypes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "list of known action_types."
        },
        "objectTypes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "list of known action_types."
        }
      },
      "title": "ListAuditEventTypesResponse lists known events and objects types"
    },
    "v1ListAuditEventsRequest": {
      "type": "object",
      "properties": {
        "fsp": {
          "$ref": "#/definitions/fspFilteringSortingPagination",
          "description": "fsp used for filtering/sorting/paginating the request. Optional param."
        }
      },
      "description": "ListAuditEventsRequest to view events that were previously saved."
    },
    "v1ListAuditEventsResponse": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1AuditEvent"
          },
          "description": "list of events."
        },
        "pageTotals": {
          "$ref": "#/definitions/fspPageTotals",
          "description": "pagination details."
        }
      },
      "description": "ListAuditEventsResponse lists stored events."
    }
  }
}